Add a note to login with email, not username
Because I have usernames "saved" to my http://localhost:3000/ I wasn't able to see that the login form was asking for an *email* and *password* - not a username.

This sent me on a wild goose chase for a while.  So I decided to add a pull request with an appropriate note to help any future travelers like me down this path.

Here are some forum links where this issue is discussed:
http://forums.orionjs.org/t/microscope-example-fails-with-login-forbidden-error/241/3
http://forums.orionjs.org/t/solved-microscope-example-login-forbidden-failure/167

Please consider adding this to the tutorial.  Thanks.