Added clearity for Deny Callbacks in posts.js
I'm new to Meteor Development, and OrionJS for that matter.  This little problem had me running around for hours to figure out.  I thought that maybe we could solve this for future readers by documenting that we have to remove this Deny Callback in order to make any updates from the admin panel.